feat(ai-tools): extract web lookup core
